+++ beets/ui/__init__.py
@@ -703,7 +703,6 @@ def main(args=None, configfh=None):
     # Load requested plugins.
     plugnames = config_val(config, 'beets', 'plugins', '')
     plugins.load_plugins(plugnames.split())
-    plugins.load_listeners()
     plugins.send("pluginload")
     plugins.configure(config)