+++ test/test_art.py
@@ -1,5 +1,5 @@
 # This file is part of beets.
-# Copyright 2011, Adrian Sampson.
+# Copyright 2012, Adrian Sampson.
 #
 # Permission is hereby granted, free of charge, to any person obtaining
 # a copy of this software and associated documentation files (the
@@ -16,8 +16,10 @@
 
 import _common
 from _common import unittest
-from beets.autotag import art
+from beetsplug import fetchart
 from beets.autotag import AlbumInfo
+from beets import library
+from beets import importer
 import os
 import shutil
 import StringIO
@@ -38,13 +40,13 @@ class MockUrlRetrieve(object):
 
 class AmazonArtTest(unittest.TestCase):
     def test_invalid_type_returns_none(self):
-        art.urllib.urlretrieve = MockUrlRetrieve('path', '')
-        artpath = art.art_for_asin('xxxx')
+        fetchart.urllib.urlretrieve = MockUrlRetrieve('path', '')
+        artpath = fetchart.art_for_asin('xxxx')
         self.assertEqual(artpath, None)
 
     def test_jpeg_type_returns_path(self):
-        art.urllib.urlretrieve = MockUrlRetrieve('somepath', 'image/jpeg')
-        artpath = art.art_for_asin('xxxx')
+        fetchart.urllib.urlretrieve = MockUrlRetrieve('somepath', 'image/jpeg')
+        artpath = fetchart.art_for_asin('xxxx')
         self.assertEqual(artpath, 'somepath')
 
 class FSArtTest(unittest.TestCase):
@@ -56,84 +58,116 @@ class FSArtTest(unittest.TestCase):
 
     def test_finds_jpg_in_directory(self):
         _common.touch(os.path.join(self.dpath, 'a.jpg'))
-        fn = art.art_in_path(self.dpath)
+        fn = fetchart.art_in_path(self.dpath)
         self.assertEqual(fn, os.path.join(self.dpath, 'a.jpg'))
 
     def test_appropriately_named_file_takes_precedence(self):
         _common.touch(os.path.join(self.dpath, 'a.jpg'))
         _common.touch(os.path.join(self.dpath, 'cover.jpg'))
-        fn = art.art_in_path(self.dpath)
+        fn = fetchart.art_in_path(self.dpath)
         self.assertEqual(fn, os.path.join(self.dpath, 'cover.jpg'))
 
     def test_non_image_file_not_identified(self):
         _common.touch(os.path.join(self.dpath, 'a.txt'))
-        fn = art.art_in_path(self.dpath)
+        fn = fetchart.art_in_path(self.dpath)
         self.assertEqual(fn, None)
 
 class CombinedTest(unittest.TestCase):
     def setUp(self):
         self.dpath = os.path.join(_common.RSRC, 'arttest')
         os.mkdir(self.dpath)
-        self.old_urlopen = art.urllib.urlopen
-        art.urllib.urlopen = self._urlopen
+        self.old_urlopen = fetchart.urllib.urlopen
+        fetchart.urllib.urlopen = self._urlopen
         self.page_text = ""
+        self.urlopen_called = False
     def tearDown(self):
         shutil.rmtree(self.dpath)
-        art.urllib.urlopen = self.old_urlopen
+        fetchart.urllib.urlopen = self.old_urlopen
 
     def _urlopen(self, url):
         self.urlopen_called = True
+        self.fetched_url = url
         return StringIO.StringIO(self.page_text)
 
     def test_main_interface_returns_amazon_art(self):
-        art.urllib.urlretrieve = MockUrlRetrieve('anotherpath', 'image/jpeg')
-        album = AlbumInfo(None, None, None, None, None, asin='xxxx')
-        artpath = art.art_for_album(album, None)
+        fetchart.urllib.urlretrieve = \
+                MockUrlRetrieve('anotherpath', 'image/jpeg')
+        album = _common.Bag(asin='xxxx')
+        artpath = fetchart.art_for_album(album, None)
         self.assertEqual(artpath, 'anotherpath')
 
     def test_main_interface_returns_none_for_missing_asin_and_path(self):
-        album = AlbumInfo(None, None, None, None, None, asin=None)
-        artpath = art.art_for_album(album, None)
+        album = _common.Bag()
+        artpath = fetchart.art_for_album(album, None)
         self.assertEqual(artpath, None)
 
     def test_main_interface_gives_precedence_to_fs_art(self):
         _common.touch(os.path.join(self.dpath, 'a.jpg'))
-        art.urllib.urlretrieve = MockUrlRetrieve('anotherpath', 'image/jpeg')
-        album = AlbumInfo(None, None, None, None, None, asin='xxxx')
-        artpath = art.art_for_album(album, self.dpath)
+        fetchart.urllib.urlretrieve = \
+                MockUrlRetrieve('anotherpath', 'image/jpeg')
+        album = _common.Bag(asin='xxxx')
+        artpath = fetchart.art_for_album(album, self.dpath)
         self.assertEqual(artpath, os.path.join(self.dpath, 'a.jpg'))
 
     def test_main_interface_falls_back_to_amazon(self):
-        art.urllib.urlretrieve = MockUrlRetrieve('anotherpath', 'image/jpeg')
-        album = AlbumInfo(None, None, None, None, None, asin='xxxx')
-        artpath = art.art_for_album(album, self.dpath)
+        fetchart.urllib.urlretrieve = \
+                MockUrlRetrieve('anotherpath', 'image/jpeg')
+        album = _common.Bag(asin='xxxx')
+        artpath = fetchart.art_for_album(album, self.dpath)
         self.assertEqual(artpath, 'anotherpath')
 
     def test_main_interface_tries_amazon_before_aao(self):
-        self.urlopen_called = False
-        art.urllib.urlretrieve = MockUrlRetrieve('anotherpath', 'image/jpeg')
-        album = AlbumInfo(None, None, None, None, None, asin='xxxx')
-        art.art_for_album(album, self.dpath)
+        fetchart.urllib.urlretrieve = \
+                MockUrlRetrieve('anotherpath', 'image/jpeg')
+        album = _common.Bag(asin='xxxx')
+        fetchart.art_for_album(album, self.dpath)
         self.assertFalse(self.urlopen_called)
 
     def test_main_interface_falls_back_to_aao(self):
-        self.urlopen_called = False
-        art.urllib.urlretrieve = MockUrlRetrieve('anotherpath', 'text/html')
-        album = AlbumInfo(None, None, None, None, None, asin='xxxx')
-        art.art_for_album(album, self.dpath)
+        fetchart.urllib.urlretrieve = \
+                MockUrlRetrieve('anotherpath', 'text/html')
+        album = _common.Bag(asin='xxxx')
+        fetchart.art_for_album(album, self.dpath)
         self.assertTrue(self.urlopen_called)
+
+    def test_main_interface_uses_caa_when_mbid_available(self):
+        mock_retrieve = MockUrlRetrieve('anotherpath', 'image/jpeg')
+        fetchart.urllib.urlretrieve = mock_retrieve
+        album = _common.Bag(mb_albumid='releaseid', asin='xxxx')
+        artpath = fetchart.art_for_album(album, None)
+        self.assertEqual(artpath, 'anotherpath')
+        self.assertTrue('coverartarchive.org' in mock_retrieve.fetched)
+
+    def test_local_only_does_not_access_network(self):
+        mock_retrieve = MockUrlRetrieve('anotherpath', 'image/jpeg')
+        fetchart.urllib.urlretrieve = mock_retrieve
+        album = _common.Bag(mb_albumid='releaseid', asin='xxxx')
+        artpath = fetchart.art_for_album(album, self.dpath, local_only=True)
+        self.assertEqual(artpath, None)
+        self.assertFalse(self.urlopen_called)
+        self.assertFalse(mock_retrieve.fetched)
+
+    def test_local_only_gets_fs_image(self):
+        _common.touch(os.path.join(self.dpath, 'a.jpg'))
+        mock_retrieve = MockUrlRetrieve('anotherpath', 'image/jpeg')
+        fetchart.urllib.urlretrieve = mock_retrieve
+        album = _common.Bag(mb_albumid='releaseid', asin='xxxx')
+        artpath = fetchart.art_for_album(album, self.dpath, local_only=True)
+        self.assertEqual(artpath, os.path.join(self.dpath, 'a.jpg'))
+        self.assertFalse(self.urlopen_called)
+        self.assertFalse(mock_retrieve.fetched)
 
 class AAOTest(unittest.TestCase):
     def setUp(self):
-        self.old_urlopen = art.urllib.urlopen
-        self.old_urlretrieve = art.urllib.urlretrieve
-        art.urllib.urlopen = self._urlopen
+        self.old_urlopen = fetchart.urllib.urlopen
+        self.old_urlretrieve = fetchart.urllib.urlretrieve
+        fetchart.urllib.urlopen = self._urlopen
         self.retriever = MockUrlRetrieve('somepath', 'image/jpeg')
-        art.urllib.urlretrieve = self.retriever
+        fetchart.urllib.urlretrieve = self.retriever
         self.page_text = ''
     def tearDown(self):
-        art.urllib.urlopen = self.old_urlopen
-        art.urllib.urlretrieve = self.old_urlretrieve
+        fetchart.urllib.urlopen = self.old_urlopen
+        fetchart.urllib.urlretrieve = self.old_urlretrieve
 
     def _urlopen(self, url):
         return StringIO.StringIO(self.page_text)
@@ -144,15 +178,117 @@ class AAOTest(unittest.TestCase):
         <a href="TARGET_URL" title="View larger image" class="thickbox" style="color: #7E9DA2; text-decoration:none;">
         <img src="http://www.albumart.org/images/zoom-icon.jpg" alt="View larger image" width="17" height="15"  border="0"/></a>
         """
-        res = art.aao_art('x')
+        res = fetchart.aao_art('x')
         self.assertEqual(self.retriever.fetched, 'TARGET_URL')
         self.assertEqual(res, 'somepath')
 
     def test_aao_scraper_returns_none_when_no_image_present(self):
         self.page_text = "blah blah"
-        res = art.aao_art('x')
+        res = fetchart.aao_art('x')
         self.assertEqual(self.retriever.fetched, None)
         self.assertEqual(res, None)
+
+class ArtImporterTest(unittest.TestCase, _common.ExtraAsserts):
+    def setUp(self):
+        # Mock the album art fetcher to always return our test file.
+        self.art_file = os.path.join(_common.RSRC, 'tmpcover.jpg')
+        _common.touch(self.art_file)
+        self.old_afa = fetchart.art_for_album
+        self.afa_response = self.art_file
+        def art_for_album(i, p, local_only=False):
+            return self.afa_response
+        fetchart.art_for_album = art_for_album
+
+        # Test library.
+        self.libpath = os.path.join(_common.RSRC, 'tmplib.blb')
+        self.libdir = os.path.join(_common.RSRC, 'tmplib')
+        os.mkdir(self.libdir)
+        os.mkdir(os.path.join(self.libdir, 'album'))
+        itempath = os.path.join(self.libdir, 'album', 'test.mp3')
+        shutil.copyfile(os.path.join(_common.RSRC, 'full.mp3'), itempath)
+        self.lib = library.Library(self.libpath)
+        self.i = _common.item()
+        self.i.path = itempath
+        self.album = self.lib.add_album([self.i])
+        self.lib._connection().commit()
+
+        # The plugin and import configuration.
+        self.plugin = fetchart.FetchArtPlugin()
+        self.config = _common.iconfig(self.lib)
+
+        # Import task for the coroutine.
+        self.task = importer.ImportTask(None, None, [self.i])
+        self.task.is_album = True
+        self.task.album_id = self.album.id
+        info = AlbumInfo(
+            album = 'some album',
+            album_id = 'albumid',
+            artist = 'some artist',
+            artist_id = 'artistid',
+            tracks = [],
+        )
+        self.task.set_choice((info, [self.i]))
+
+    def tearDown(self):
+        fetchart.art_for_album = self.old_afa
+        if os.path.exists(self.art_file):
+            os.remove(self.art_file)
+        if os.path.exists(self.libpath):
+            os.remove(self.libpath)
+        if os.path.exists(self.libdir):
+            shutil.rmtree(self.libdir)
+
+    def _fetch_art(self, should_exist):
+        """Execute the fetch_art coroutine for the task and return the
+        album's resulting artpath. ``should_exist`` specifies whether to
+        assert that art path was set (to the correct value) or or that
+        the path was not set.
+        """
+        # Execute the two relevant parts of the importer.
+        self.plugin.fetch_art(self.config, self.task)
+        self.plugin.assign_art(self.config, self.task)
+
+        artpath = self.lib.albums()[0].artpath
+        if should_exist:
+            self.assertEqual(artpath,
+                os.path.join(os.path.dirname(self.i.path), 'cover.jpg'))
+            self.assertExists(artpath)
+        else:
+            self.assertEqual(artpath, None)
+        return artpath
+
+    def test_fetch_art(self):
+        assert not self.lib.albums()[0].artpath
+        self._fetch_art(True)
+
+    def test_art_not_found(self):
+        self.afa_response = None
+        self._fetch_art(False)
+
+    def test_no_art_for_singleton(self):
+        self.task.is_album = False
+        self._fetch_art(False)
+
+    def test_leave_original_file_in_place(self):
+        self._fetch_art(True)
+        self.assertExists(self.art_file)
+
+    def test_delete_original_file(self):
+        self.config.delete = True
+        self._fetch_art(True)
+        self.assertNotExists(self.art_file)
+
+    def test_move_original_file(self):
+        self.config.move = True
+        self._fetch_art(True)
+        self.assertNotExists(self.art_file)
+
+    def test_do_not_delete_original_if_already_in_place(self):
+        artdest = os.path.join(os.path.dirname(self.i.path), 'cover.jpg')
+        shutil.copyfile(self.art_file, artdest)
+        self.afa_response = artdest
+        self._fetch_art(True)
+
 
 def suite():
     return unittest.TestLoader().loadTestsFromName(__name__)